Update Skypoint and fix coordinate converssion problems
SkyPoint is updated to provide a catalogueCoord function that converts
'JNow' positions to J2000 correctly.
Also SkyPoint uses libnova for nutation and aberration conversions
because the SkyPoint ones were not giving correct results
Add skypoint tests for this.
Use the new catalogueCoord function for conversions from 'JNOW' to
J2000.
This gives align errors using the CCD and telescope simulator of zero to
1 arc second.